fix: wrong class edit
