+++ src/main/java/com/devoxx/genie/chatmodel/openai/OpenAIChatModelFactory.java
@@ -5,6 +5,7 @@ import com.devoxx.genie.model.ChatModel;
 import com.devoxx.genie.model.LanguageModel;
 import com.devoxx.genie.model.enumarations.ModelProvider;
 import com.devoxx.genie.service.DevoxxGenieSettingsServiceProvider;
+import com.jgoodies.common.base.Strings;
 import dev.langchain4j.model.chat.ChatLanguageModel;
 import dev.langchain4j.model.chat.StreamingChatLanguageModel;
 import dev.langchain4j.model.openai.OpenAiChatModel;
@@ -19,26 +20,36 @@ public class OpenAIChatModelFactory implements ChatModelFactory {
     @Override
     public ChatLanguageModel createChatModel(@NotNull ChatModel chatModel) {
         boolean isO1 = chatModel.getModelName().startsWith("o1-");
-        return OpenAiChatModel.builder()
+
+        final var builder = OpenAiChatModel.builder()
                 .apiKey(getApiKey())
                 .modelName(chatModel.getModelName())
                 .maxRetries(chatModel.getMaxRetries())
                 .temperature(isO1 ? 1.0 : chatModel.getTemperature())
                 .timeout(Duration.ofSeconds(chatModel.getTimeout()))
-                .topP(isO1 ? 1.0 : chatModel.getTopP())
-                .build();
+                .topP(isO1 ? 1.0 : chatModel.getTopP());
+
+        if (Strings.isNotBlank(DevoxxGenieSettingsServiceProvider.getInstance().getCustomOpenAIUrl())) {
+            builder.baseUrl(DevoxxGenieSettingsServiceProvider.getInstance().getCustomOpenAIUrl());
+        }
+
+        return builder.build();
     }
 
     @Override
     public StreamingChatLanguageModel createStreamingChatModel(@NotNull ChatModel chatModel) {
         boolean isO1 = chatModel.getModelName().startsWith("o1-");
-        return OpenAiStreamingChatModel.builder()
+        final var builder = OpenAiStreamingChatModel.builder()
                 .apiKey(getApiKey())
                 .modelName(chatModel.getModelName())
                 .temperature(isO1 ? 1.0 : chatModel.getTemperature())
                 .topP(isO1 ? 1.0 : chatModel.getTopP())
-                .timeout(Duration.ofSeconds(chatModel.getTimeout()))
-                .build();
+                .timeout(Duration.ofSeconds(chatModel.getTimeout()));
+
+        if (Strings.isNotBlank(DevoxxGenieSettingsServiceProvider.getInstance().getCustomOpenAIUrl())) {
+            builder.baseUrl(DevoxxGenieSettingsServiceProvider.getInstance().getCustomOpenAIUrl());
+        }
+        return builder.build();
     }
 
     @Override
